cursor appears on top of player

+++ src/tilemap.rs
@@ -1,4 +1,4 @@
-use tinybit::{Color, WorldPos, ScreenSize};
+use tinybit::{Color, ScreenSize, WorldPos};
 
 pub type TilemapMeh = Tilemap<ThrowAwayThisProvider>;
 
@@ -7,14 +7,15 @@ pub struct ThrowAwayThisProvider;
 impl MapSource for ThrowAwayThisProvider {
     fn get_region(&self, pos: WorldPos, size: ScreenSize) -> Vec<Tile> {
         let mut tiles = Vec::new();
-        let start_x = pos.x - size.width as isize / 2;
-        let start_y = pos.y - size.height as isize / 2;
-        let end_x = pos.x + size.width as isize/ 2;
-        let end_y = pos.y + size.height as isize/ 2;
+
+        let start_x = (pos.x - size.width as f32 / 2.0) as isize;
+        let start_y = (pos.y - size.height as f32 / 2.0) as isize;
+        let end_x   = (pos.x + size.width as f32 / 2.0) as isize;
+        let end_y   = (pos.y + size.height as f32 / 2.0) as isize;
 
         for x in start_x..=end_x {
             for y in start_y..=end_y {
-                tiles.push(Tile::new('.', WorldPos::new(x, y), None));
+                tiles.push(Tile::new('.', WorldPos::new(x as f32, y as f32), None));
             }
         }
 
@@ -34,7 +35,6 @@ impl Tile {
         Self { glyph, pos, color }
     }
 }
-
 
 trait MapSource {
     fn get_region(&self, pos: WorldPos, size: ScreenSize) -> Vec<Tile>;
@@ -61,7 +61,7 @@ impl<T: MapSource> Tilemap<T> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use tinybit::{Pixel, WorldPos, ScreenSize};
+    use tinybit::{Pixel, ScreenSize, WorldPos};
 
     struct TestProvider;
 
@@ -70,8 +70,8 @@ mod test {
             let mut tiles = Vec::new();
             let start_x = pos.x - size.width as isize / 2;
             let start_y = pos.x - size.width as isize / 2;
-            let end_x = pos.x + size.width as isize/ 2;
-            let end_y = pos.x + size.width as isize/ 2;
+            let end_x = pos.x + size.width as isize / 2;
+            let end_y = pos.x + size.width as isize / 2;
 
             for x in start_x..=end_x {
                 for y in start_y..=end_y {
@@ -91,6 +91,5 @@ mod test {
             .get_region(WorldPos::zero(), ScreenSize::new(10, 10));
 
         assert_eq!(region[0], Tile::new('#', WorldPos::new(-5, -5), None));
-
     }
 }
